postgres_to_es/requirements/requirements.txt
psycopg==3.2.4
python-dotenv==1.0.1
elasticsearch==8.17.2
redis==5.2.1